Validate the arguments passed to public functions.
In order to behave nicely even when the user screws up big time, check
if "mb_param" and "mb_param->obj" are not NULL. Convenience macros have
been defined for this.

Also, the function "set_message_length_tcp" has been made "static"
because it doesn't have a prototype in the header file.