src/version.h.in: Add versioning infrastructure.
The new file <modbus/version.h> provides macros to determine the library
version at compile time and functions to determine the version at
runtime.

Using the macros, a user can determine whether or not a given feature is
present in the library or not by checking the version number. For
example:

  #if MB_VERSION < MB_VERSION_ENCODE(2,1,0)
  /* compatibility code for old version */
  float modbus_read_float(const uint16_t *src) { ... }
  #endif

Using the functions, a user can check the version at runtime. For example:

  if (modbus_version () == MB_VERSION)
    puts ("This is $my_great_program "
        "using libmodbus version "MB_VERSION_STRING"\n");
  else
    printf ("This is $my_great_program "
        "using libmodbus version %s "
        "(compiled with version "MB_VERSION_STRING")\n",
        modbus_version_string ());

The package version has been updated to 2.1.0 and the SONAME version has
been updated to 3.0.0. Resolves #549410.